docs/requirements-pdf.txt
pypdf>=3.0.0
Pillow>=10.0.0
pandas>=2.2.2
openpyxl>=3.1.2
